+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to
 #### Breaking Changes
 - bpf: test gcc toolchain;
 #### Added
+- shell: test `mapfile` builtin;
 - scripts: add libcpu.sh;
 - top-Makefile: script support V=1,DEBUG=1;
 - cxl/fm: add mctp subdir;